Integrates run_dbt task into update_socrata_table TaskGroup.
