+++ airflow/dags/tasks/socrata_tasks.py
@@ -2,6 +2,7 @@ import datetime as dt
 from logging import Logger
 import os
 from pathlib import Path
+import subprocess
 from urllib.request import urlretrieve
 
 from airflow.decorators import task, task_group
@@ -696,19 +697,17 @@ def dbt_clean_model_ready(task_logger: Logger, **kwargs) -> str:
         "intermediate",
         f"{socrata_metadata.table_name}_clean.sql",
     )
-    host_file_path = str(file_path).replace(airflow_home, "/airflow")
     if file_path.is_file():
         task_logger.info(f"Found a _clean stage dbt model that looks finished; Ending")
-        return "update_socrata_table.endpoint"
+        return "update_socrata_table.run_dbt_models__standardized_onward"
     else:
         return "update_socrata_table.dbt_make_clean_model"
 
 
 @task(trigger_rule=TriggerRule.NONE_FAILED_MIN_ONE_SUCCESS)
-def dbt_make_clean_model(conn_id: str, task_logger: Logger, **kwargs) -> SocrataTableMetadata:
+def dbt_make_clean_model(task_logger: Logger, **kwargs) -> SocrataTableMetadata:
     ti = kwargs["ti"]
     socrata_metadata = ti.xcom_pull(task_ids="update_socrata_table.download_fresh_data")
-    engine = get_pg_engine(conn_id=conn_id)
     airflow_home = os.environ["AIRFLOW_HOME"]
     clean_file_path = Path(airflow_home).joinpath(
         "dbt",
@@ -732,6 +731,20 @@ def highlight_unfinished_dbt_standardized_stub(task_logger: Logger) -> str:
         + "(REPLACE_WITH_COMPOSITE_KEY_COLUMNS or REPLACE_WITH_BETTER_id)."
     )
     return "Please and thank you!"
+
+
+@task(trigger_rule=TriggerRule.NONE_FAILED_MIN_ONE_SUCCESS)
+def run_dbt_models__standardized_onward(task_logger: Logger, **kwargs) -> SocrataTableMetadata:
+    ti = kwargs["ti"]
+    socrata_metadata = ti.xcom_pull(task_ids="update_socrata_table.download_fresh_data")
+    dbt_cmd = f"""cd /opt/airflow/dbt && \
+                  dbt --warn-error run --select \
+                  re_dbt.intermediate.{socrata_metadata.table_name}_standardized+"""
+    task_logger.info(f"dbt run command: {dbt_cmd}")
+    subproc_output = subprocess.run(dbt_cmd, shell=True, capture_output=True, text=True)
+    for el in subproc_output.stdout.split("\n"):
+        task_logger.info(f"{el}")
+    return socrata_metadata
 
 
 @task(trigger_rule=TriggerRule.NONE_FAILED_MIN_ONE_SUCCESS)
@@ -770,7 +783,8 @@ def update_socrata_table(
         task_logger=task_logger,
     )
     clean_model_ready_1 = dbt_clean_model_ready(task_logger=task_logger)
-    make_clean_model_1 = dbt_make_clean_model(conn_id=conn_id, task_logger=task_logger)
+    make_clean_model_1 = dbt_make_clean_model(task_logger=task_logger)
+    run_dbt_models_1 = run_dbt_models__standardized_onward(task_logger=task_logger)
     endpoint_1 = endpoint(task_logger=task_logger)
     update_metadata_false_1 = update_result_of_check_in_metadata_table(
         conn_id=conn_id, task_logger=task_logger, data_updated=False
@@ -784,20 +798,26 @@ def update_socrata_table(
         extract_data_1,
         load_data_tg_1,
         std_model_ready_1,
+    )
+    chain(
+        std_model_ready_1,
         [
             Label("No dbt _standardized model found"),
             Label("dbt _standardized model needs review"),
-            Label("dbt _standardized model looks good"),
         ],
         [
             make_standardized_stage_1,
             std_model_unfinished_1,
-            clean_model_ready_1,
         ],
         endpoint_1,
     )
     chain(
-        clean_model_ready_1, [Label("dbt _clean model looks good!"), make_clean_model_1], endpoint_1
+        std_model_ready_1,
+        Label("dbt _standardized model looks good"),
+        clean_model_ready_1,
+        [Label("dbt _clean model looks good!"), make_clean_model_1],
+        run_dbt_models_1,
+        endpoint_1,
     )
     chain(
         metadata_1,
